requirements.txt
python-telegram-bot
groq
requests
Pillow